Fix type conversion & comparison warnings
Each too small to put into a separate commit. Used best judgement between
simply casting to the desired type and changing the desired type to better
match inputs and outputs. EG when reading a size_t into a local variable
as a uint32_t, only to write out to a size_t is best fixed by changing
the local variable to size_t instead of narrowing from size_t to
uint32_t.

+++ framework/decode/common_handle_mapping_util.h
@@ -214,7 +214,7 @@ static void AddHandleArrayAsync(format::HandleId        parent_id,
             initial_info.capture_id          = ids[i];
             initial_info.parent_id           = parent_id;
             initial_info.future              = future;
-            initial_info.future_handle_index = i;
+            initial_info.future_handle_index = static_cast<uint32_t>(i);
             (object_info_table->*AddFunc)(std::move(initial_info));
         }
     }
